requirements.txt
colorama==0.4.6
mysql-connector-python==8.2.0
protobuf==4.21.12
